requirements.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.2
click==7.1.2
Flask==1.1.2
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.8
MarkupSafe==1.1.1
numpy==1.18.3
opencv-python==4.2.0.34
pytz==2020.1
vine==1.3.0
Werkzeug==1.0.1
zipp==3.1.0
python-dotenv~=0.13.0
setuptools~=46.4.0
boto3~=1.13.12
pandas~=1.0.3
mysql-connector-python~=8.0.20
termcolor~=1.1.0
psycopg2~=2.8.5
requests~=2.23.0